got sign in/out working for authentication web service
git-svn-id: https://svn.code.sf.net/p/phpscheduleit/code/development@1751 12d54735-0ee2-4768-9b18-c71e5e561e4c

+++ lib/Application/Authentication/WebServiceAuthentication.php
@@ -39,10 +39,11 @@ interface IWebServiceAuthentication
 	public function Login($username);
 
 	/**
+	 * @param string $publicUserId
 	 * @param string $sessionToken
 	 * @return void
 	 */
-	public function Logout($sessionToken);
+	public function Logout($publicUserId, $sessionToken);
 }
 
 class WebServiceAuthentication implements IWebServiceAuthentication
@@ -92,15 +93,16 @@ class WebServiceAuthentication implements IWebServiceAuthentication
 	}
 
 	/**
+	 * @param string $publicUserId
 	 * @param string $sessionToken
 	 * @return void
 	 */
-	public function Logout($sessionToken)
+	public function Logout($publicUserId, $sessionToken)
 	{
 		Log::Debug('Logout sessionToken: %s', $sessionToken);
 
 		$userSession = $this->userSessionRepository->LoadBySessionToken($sessionToken);
-		if ($userSession != null)
+		if ($userSession != null && $userSession->PublicId == $publicUserId)
 		{
 			$this->userSessionRepository->Delete($userSession);
 			$this->authentication->Logout($userSession);
